Inverted the previous and next buttons
It looked to me that the buttons were inverted. It might be the case that it's a ltr-rtl issue?
I'm not sure.

However, I put the *previous post* button to the left and the *next post* button to the right.

Take a look and let me know if this fixes the issue or if it's just a misconfiguration on my end.