Improve the widget functionality of the frontpage
Now the widgets are flexible as you like them to be. The include checks if parameters are given, if not, nothing will be generated. For example: If you don't want headline, don't write them in frontmatter. If you don't like the more button (hello @fedorov ;) just don't use an url.